Refine builder
